feat: convert into a crate
