+++ Cargo.toml
@@ -1,17 +1,26 @@
 [package]
 name = "rocket-tracing-fairing-example"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
+
+[lib]
+name = "rocket_tracing_fairing"
+path = "src/lib.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 rocket = { version = "0.5.0", features = ["json"] }
-serde = {version="1.0.197", features=["derive"]}
+serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
 tracing = "0.1.40"
 tracing-appender = "0.2.3"
 tracing-log = "0.2.0"
-tracing-subscriber = {version="0.3.18", features=["env-filter", "json", "registry", "smallvec"]}
-uuid = {version="1.8.0", features = ["v4"]}
+tracing-subscriber = { version = "0.3.18", features = [
+    "env-filter",
+    "json",
+    "registry",
+    "smallvec",
+] }
+uuid = { version = "1.8.0", features = ["v4"] }
 yansi = "0.5.1"
