+++ src/lib.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+extern crate tracing;
+
+pub mod logging;
+pub mod spans;
