+++ src/logging.rs
@@ -1,0 +1,96 @@
+use tracing_subscriber::field::MakeExt;
+use tracing_subscriber::Layer;
+use tracing_subscriber::{registry::LookupSpan, EnvFilter};
+use yansi::Paint;
+
+pub enum LogType {
+    Formatted,
+    Json,
+}
+
+impl From<String> for LogType {
+    fn from(input: String) -> Self {
+        match input.as_str() {
+            "formatted" => Self::Formatted,
+            "json" => Self::Json,
+            _ => panic!("Unkown log type {}", input),
+        }
+    }
+}
+
+pub fn default_logging_layer<S>() -> impl Layer<S>
+where
+    S: tracing::Subscriber,
+    S: for<'span> LookupSpan<'span>,
+{
+    let field_format = tracing_subscriber::fmt::format::debug_fn(|writer, field, value| {
+        // We'll format the field name and value separated with a colon.
+        if field.name() == "message" {
+            write!(writer, "{:?}", Paint::new(value).bold())
+        } else {
+            write!(writer, "{}: {:?}", field, Paint::default(value).bold())
+        }
+    })
+    .delimited(", ")
+    .display_messages();
+
+    tracing_subscriber::fmt::layer()
+        .fmt_fields(field_format)
+        // Configure the formatter to use `print!` rather than
+        // `stdout().write_str(...)`, so that logs are captured by libtest's test
+        // capturing.
+        .with_test_writer()
+}
+
+pub fn json_logging_layer<
+    S: for<'a> tracing_subscriber::registry::LookupSpan<'a> + tracing::Subscriber,
+>() -> impl tracing_subscriber::Layer<S> {
+    Paint::disable();
+
+    tracing_subscriber::fmt::layer()
+        .json()
+        // Configure the formatter to use `print!` rather than
+        // `stdout().write_str(...)`, so that logs are captured by libtest's test
+        // capturing.
+        .with_test_writer()
+}
+
+#[derive(PartialEq, Eq, Debug, Clone, Copy)]
+pub enum LogLevel {
+    /// Only shows errors and warnings: `"critical"`.
+    Critical,
+    /// Shows errors, warnings, and some informational messages that are likely
+    /// to be relevant when troubleshooting such as configuration: `"support"`.
+    Support,
+    /// Shows everything except debug and trace information: `"normal"`.
+    Normal,
+    /// Shows everything: `"debug"`.
+    Debug,
+    /// Shows nothing: "`"off"`".
+    Off,
+}
+
+impl From<&str> for LogLevel {
+    fn from(s: &str) -> Self {
+        match &*s.to_ascii_lowercase() {
+            "critical" => LogLevel::Critical,
+            "support" => LogLevel::Support,
+            "normal" => LogLevel::Normal,
+            "debug" => LogLevel::Debug,
+            "off" => LogLevel::Off,
+            _ => panic!("a log level (off, debug, normal, support, critical)"),
+        }
+    }
+}
+
+pub fn filter_layer(level: LogLevel) -> EnvFilter {
+    let filter_str = match level {
+        LogLevel::Critical => "warn,hyper=off,rustls=off",
+        LogLevel::Support => "warn,rocket::support=info,hyper=off,rustls=off",
+        LogLevel::Normal => "info,hyper=off,rustls=off",
+        LogLevel::Debug => "trace",
+        LogLevel::Off => "off",
+    };
+
+    tracing_subscriber::filter::EnvFilter::try_new(filter_str).expect("filter string must parse")
+}
